Added better CURRENT file handling
Normal databases store the name of the manifest file in the CURRENT file.
Marketplace worlds store some binary data in this file that seems to always start with a null character.
The old code has been modified to extract the string up to a non-printable character.
If the final result is an empty string a not supported status is returned.