Lock the current file while the db is open
Minecraft does not seem to care about the LOCK file and will just open the database even if we have it open.
The only way I can find to stop it is by locking the CURRENT file.
When modifying the CURRENT file we need to unlock it before and re-lock it after.